+++ Cesium3DTilesContent/src/BinaryToGltfConverter.cpp
@@ -13,30 +13,29 @@
 namespace Cesium3DTilesContent {
 CesiumGltfReader::GltfReader BinaryToGltfConverter::_gltfReader;
 
-GltfConverterResult BinaryToGltfConverter::convertImmediate(
-    const std::span<const std::byte>& gltfBinary,
-    const CesiumGltfReader::GltfReaderOptions& options,
-    const AssetFetcher& assetFetcher) {
-  CesiumGltfReader::GltfReaderResult loadedGltf =
-      _gltfReader.readGltf(gltfBinary, options);
-
-  if (loadedGltf.model) {
-    loadedGltf.model->extras["gltfUpAxis"] =
-        static_cast<std::underlying_type_t<CesiumGeometry::Axis>>(
-            assetFetcher.upAxis);
-  }
-  GltfConverterResult result;
-  result.model = std::move(loadedGltf.model);
-  result.errors.errors = std::move(loadedGltf.errors);
-  result.errors.warnings = std::move(loadedGltf.warnings);
-  return result;
-}
-
 CesiumAsync::Future<GltfConverterResult> BinaryToGltfConverter::convert(
     const std::span<const std::byte>& gltfBinary,
     const CesiumGltfReader::GltfReaderOptions& options,
     const AssetFetcher& assetFetcher) {
-  return assetFetcher.asyncSystem.createResolvedFuture(
-      convertImmediate(gltfBinary, options, assetFetcher));
+  return BinaryToGltfConverter::_gltfReader
+      .readGltfAndExternalData(
+          gltfBinary,
+          assetFetcher.asyncSystem,
+          assetFetcher.requestHeaders,
+          assetFetcher.pAssetAccessor,
+          assetFetcher.baseUrl,
+          options)
+      .thenInWorkerThread([upAxis = assetFetcher.upAxis](
+                              CesiumGltfReader::GltfReaderResult&& gltfResult) {
+        if (gltfResult.model) {
+          gltfResult.model->extras["gltfUpAxis"] =
+              static_cast<std::underlying_type_t<CesiumGeometry::Axis>>(upAxis);
+        }
+        GltfConverterResult result;
+        result.model = std::move(gltfResult.model);
+        result.errors.errors = std::move(gltfResult.errors);
+        result.errors.warnings = std::move(gltfResult.warnings);
+        return result;
+      });
 }
 } // namespace Cesium3DTilesContent